[All] Cleanup code in order to support certification for Win8. Still need to remove FileStream/File access and minor references.

+++ Source/SharpDX/DataBuffer.cs
@@ -631,7 +631,7 @@ namespace SharpDX
         {
             unsafe
             {
-                Utilities.CopyMemory((IntPtr)(_buffer + positionInBytes), source, (ulong)count);
+                Utilities.CopyMemory((IntPtr)(_buffer + positionInBytes), source, (int)count);
             }
         }
